classification/requirements.txt
grpcio-tools
scikit-learn==1.0.2